# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: AdventureX Cloud Media Service
  summary: User-owned audio ingestion and deterministic NFC media playback.
  description: Authenticate users and fixed devices, normalize uploaded audio, render the Sound Visualization,
    publish immutable H.264/MP3 media, and serve NFC playback.
  version: 2.0.0
  x-service-id: advx26-cloud-media
servers:
  - url: http://127.0.0.1:9000
    description: Configured public Cloud Media Service endpoint.
paths:
  /c/{content_id}:
    get:
      tags:
        - devices
      summary: Resolve NFC content for Trigger
      description: Return the complete READY Compact Content document for the fixed Trigger board.
      operationId: resolveTriggerContent
      security:
        - TriggerToken: []
      parameters:
        - name: content_id
          in: path
          required: true
          schema:
            type: string
            title: Content Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompactContent'
        '401':
          description: A required Bearer Token is missing or invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: The authenticated role is not allowed to use this operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: The content or media object does not exist or is not visible to this role.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: A required service dependency or immutable object is unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/contents/{content_id}/assets/{asset_kind}:
    get:
      tags:
        - devices
      summary: Download immutable playback asset
      description: Playback-only GET response with Range and If-Range support.
      operationId: getPlaybackAsset
      security:
        - PlaybackToken: []
      parameters:
        - name: content_id
          in: path
          required: true
          schema:
            type: string
            title: Content Id
        - name: asset_kind
          in: path
          required: true
          schema:
            type: string
            enum:
              - video
              - audio
              - audio-index
          description: 'Immutable asset type: H.264 MP4, normalized MP3, or AIX1 index.'
      responses:
        '200':
          description: Complete immutable media object.
          headers:
            Accept-Ranges:
              description: Indicates byte range support.
              schema:
                type: string
                const: bytes
            Content-Length:
              description: Exact number of bytes in this response body.
              schema:
                type: integer
                minimum: 0
            ETag:
              description: Strong SHA-256-derived entity tag.
              schema:
                type: string
            Cache-Control:
              description: Private immutable cache policy for READY media.
              schema:
                type: string
                const: private, max-age=31536000, immutable
            Vary:
              description: The representation varies by Authorization.
              schema:
                type: string
                const: Authorization
          content:
            video/mp4:
              schema:
                type: string
                format: binary
            audio/mpeg:
              schema:
                type: string
                format: binary
            application/octet-stream:
              schema:
                type: string
                format: binary
        '206':
          description: One satisfiable byte range.
          headers:
            Accept-Ranges:
              description: Indicates byte range support.
              schema:
                type: string
                const: bytes
            Content-Length:
              description: Exact number of bytes in this response body.
              schema:
                type: integer
                minimum: 0
            ETag:
              description: Strong SHA-256-derived entity tag.
              schema:
                type: string
            Cache-Control:
              description: Private immutable cache policy for READY media.
              schema:
                type: string
                const: private, max-age=31536000, immutable
            Vary:
              description: The representation varies by Authorization.
              schema:
                type: string
                const: Authorization
            Content-Range:
              description: Returned inclusive byte range and total length.
              schema:
                type: string
                example: bytes 0-1023/4096
          content:
            video/mp4:
              schema:
                type: string
                format: binary
            audio/mpeg:
              schema:
                type: string
                format: binary
            application/octet-stream:
              schema:
                type: string
                format: binary
        '416':
          description: The requested byte range is invalid or outside the object.
          headers:
            Accept-Ranges:
              description: Indicates byte range support.
              schema:
                type: string
                const: bytes
            Content-Length:
              description: Exact number of bytes in this response body.
              schema:
                type: integer
                minimum: 0
            ETag:
              description: Strong SHA-256-derived entity tag.
              schema:
                type: string
            Cache-Control:
              description: Private immutable cache policy for READY media.
              schema:
                type: string
                const: private, max-age=31536000, immutable
            Vary:
              description: The representation varies by Authorization.
              schema:
                type: string
                const: Authorization
            Content-Range:
              description: Unsatisfied range with the current object length.
              schema:
                type: string
                example: bytes */4096
        '401':
          description: A required Bearer Token is missing or invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: The authenticated role is not allowed to use this operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: The content or media object does not exist or is not visible to this role.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: A required service dependency or immutable object is unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    head:
      tags:
        - devices
      summary: Inspect immutable playback asset
      description: Playback-only HEAD response with Range and If-Range support.
      operationId: headPlaybackAsset
      security:
        - PlaybackToken: []
      parameters:
        - name: content_id
          in: path
          required: true
          schema:
            type: string
            title: Content Id
        - name: asset_kind
          in: path
          required: true
          schema:
            type: string
            enum:
              - video
              - audio
              - audio-index
          description: 'Immutable asset type: H.264 MP4, normalized MP3, or AIX1 index.'
      responses:
        '200':
          description: Complete immutable media object.
          headers:
            Accept-Ranges:
              description: Indicates byte range support.
              schema:
                type: string
                const: bytes
            Content-Length:
              description: Exact number of bytes in this response body.
              schema:
                type: integer
                minimum: 0
            ETag:
              description: Strong SHA-256-derived entity tag.
              schema:
                type: string
            Cache-Control:
              description: Private immutable cache policy for READY media.
              schema:
                type: string
                const: private, max-age=31536000, immutable
            Vary:
              description: The representation varies by Authorization.
              schema:
                type: string
                const: Authorization
        '206':
          description: One satisfiable byte range.
          headers:
            Accept-Ranges:
              description: Indicates byte range support.
              schema:
                type: string
                const: bytes
            Content-Length:
              description: Exact number of bytes in this response body.
              schema:
                type: integer
                minimum: 0
            ETag:
              description: Strong SHA-256-derived entity tag.
              schema:
                type: string
            Cache-Control:
              description: Private immutable cache policy for READY media.
              schema:
                type: string
                const: private, max-age=31536000, immutable
            Vary:
              description: The representation varies by Authorization.
              schema:
                type: string
                const: Authorization
            Content-Range:
              description: Returned inclusive byte range and total length.
              schema:
                type: string
                example: bytes 0-1023/4096
        '416':
          description: The requested byte range is invalid or outside the object.
          headers:
            Accept-Ranges:
              description: Indicates byte range support.
              schema:
                type: string
                const: bytes
            Content-Length:
              description: Exact number of bytes in this response body.
              schema:
                type: integer
                minimum: 0
            ETag:
              description: Strong SHA-256-derived entity tag.
              schema:
                type: string
            Cache-Control:
              description: Private immutable cache policy for READY media.
              schema:
                type: string
                const: private, max-age=31536000, immutable
            Vary:
              description: The representation varies by Authorization.
              schema:
                type: string
                const: Authorization
            Content-Range:
              description: Unsatisfied range with the current object length.
              schema:
                type: string
                example: bytes */4096
        '401':
          description: A required Bearer Token is missing or invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: The authenticated role is not allowed to use this operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: The content or media object does not exist or is not visible to this role.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: A required service dependency or immutable object is unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CompactContent:
      properties:
        schema_version:
          type: integer
          const: 1
          title: Schema Version
        content_id:
          type: string
          pattern: ^[0-9a-f]{32}$
          title: Content Id
        state:
          type: string
          const: READY
          title: State
        duration_ms:
          type: integer
          maximum: 30000.0
          minimum: 1.0
          title: Duration Ms
        trigger:
          $ref: '#/components/schemas/TriggerPresentation'
        playback:
          $ref: '#/components/schemas/PlaybackDescriptor'
      type: object
      required:
        - schema_version
        - content_id
        - state
        - duration_ms
        - trigger
        - playback
      title: CompactContent
...
    ErrorResponse:
      properties:
        detail:
          type: string
          title: Detail
          examples:
            - 无效或缺少用户 Token
      type: object
      required:
        - detail
      title: ErrorResponse
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PlaybackAudioDescriptor:
      properties:
        url:
          type: string
          title: Url
        index_url:
          type: string
          title: Index Url
        format:
          type: string
          const: mp3
          title: Format
        bitrate:
          type: integer
          const: 128000
          title: Bitrate
        sample_rate:
          type: integer
          const: 44100
          title: Sample Rate
        channels:
          type: integer
          enum:
            - 1
            - 2
          title: Channels
        byte_length:
          type: integer
          minimum: 1.0
          title: Byte Length
        sha256:
          type: string
          pattern: ^[0-9a-f]{64}$
          title: Sha256
        etag:
          type: string
          title: Etag
        index_version:
          type: integer
          const: 1
          title: Index Version
        index_byte_length:
          type: integer
          minimum: 16.0
          title: Index Byte Length
        index_sha256:
          type: string
          pattern: ^[0-9a-f]{64}$
          title: Index Sha256
        index_etag:
          type: string
          title: Index Etag
      type: object
      required:
        - url
        - index_url
        - format
        - bitrate
        - sample_rate
        - channels
        - byte_length
        - sha256
        - etag
        - index_version
        - index_byte_length
        - index_sha256
        - index_etag
      title: PlaybackAudioDescriptor
    PlaybackDescriptor:
      properties:
        profile:
          type: string
          const: t5ai-h264-mp3-v1
          title: Profile
        video:
          $ref: '#/components/schemas/PlaybackVideoDescriptor'
        audio:
          $ref: '#/components/schemas/PlaybackAudioDescriptor'
      type: object
      required:
        - profile
        - video
        - audio
      title: PlaybackDescriptor
    PlaybackVideoDescriptor:
      properties:
        url:
          type: string
          title: Url
        format:
          type: string
          const: mp4
          title: Format
        codec:
          type: string
          const: h264
          title: Codec
        profile:
          type: string
          const: Constrained Baseline
          title: Profile
        pixel_format:
          type: string
          const: yuv420p
          title: Pixel Format
        width:
          type: integer
          const: 480
          title: Width
        height:
          type: integer
          const: 320
          title: Height
        fps:
          type: integer
          const: 10
          title: Fps
        duration_ms:
          type: integer
          maximum: 30100.0
          minimum: 1.0
          title: Duration Ms
        max_keyframe_interval_ms:
          type: integer
          maximum: 1000.0
          title: Max Keyframe Interval Ms
        byte_length:
          type: integer
          minimum: 1.0
          title: Byte Length
        sha256:
          type: string
          pattern: ^[0-9a-f]{64}$
          title: Sha256
        etag:
          type: string
          title: Etag
      type: object
      required:
        - url
        - format
        - codec
        - profile
        - pixel_format
        - width
        - height
        - fps
        - duration_ms
        - max_keyframe_interval_ms
        - byte_length
        - sha256
        - etag
      title: PlaybackVideoDescriptor
...
    TriggerPresentation:
      properties:
        display_label:
          type: string
          title: Display Label
        autoplay:
          type: boolean
          title: Autoplay
        end_behavior:
          type: string
          const: HOLD_LAST_FRAME
          title: End Behavior
        controls:
          items:
            type: string
            enum:
              - PLAY
              - PAUSE
              - SEEK
              - STOP
              - REPLAY
          type: array
          title: Controls
      type: object
      required:
        - display_label
        - autoplay
        - end_behavior
        - controls
      title: TriggerPresentation
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
...
  securitySchemes:
    TriggerToken:
      type: http
      description: Fixed Bearer Token provisioned to the Trigger board.
      scheme: bearer
      x-advx-role: trigger
      x-provisioning: Fixed Token provisioned to trigger-01.
    PlaybackToken:
      type: http
      description: Fixed Bearer Token provisioned to the Playback board.
      scheme: bearer
      x-advx-role: playback
      x-provisioning: Fixed Token provisioned to playback-01.
    UserToken:
      type: http
      description: Long-lived opaque Bearer Token returned once by the issuance endpoint.
      scheme: bearer
      x-advx-role: user
      x-provisioning: Issued by POST /api/v1/sessions after email/password login.
...